# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: Vook Frontend API
  version: 2.0.0
  description: Resource-oriented contract consumed by the standalone Vook frontend.
servers: [{ url: /api/v2 }]
security: [{ sessionCookie: [] }]
paths:
...
  /users:
    get: { operationId: listUsers, responses: { '200': { $ref: '#/components/responses/Object' } } }
    post: { operationId: inviteUser, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
...
  /employees:
    get: { operationId: listEmployees, parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }, { $ref: '#/components/parameters/Search' }], responses: { '200': { $ref: '#/components/responses/Object' } } }
    post: { operationId: createEmployee, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '201': { $ref: '#/components/responses/Employee' } } }
  /employees/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get: { operationId: getEmployee, responses: { '200': { $ref: '#/components/responses/Employee' } } }
    patch: { operationId: updateEmployee, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Employee' } } }
  /employees/{id}/actions:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post: { operationId: performEmployeeAction, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Employee' } } }
  /employees/{id}/account:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post: { operationId: provisionEmployeeAccount, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Employee' } } }
  /employees/import:
    post: { operationId: importEmployees, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Object' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: vook_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 100 } }
  requestBodies:
    Object: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
  responses:
...
    Object: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/ObjectEnvelope' } } } }
...
    Employee: { description: Employee, content: { application/json: { schema: { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/Employee' } } } } } }
...
  schemas:
...
    CompanySummary:
      type: object
      required: [id, name, companyCode]
      properties: { id: { type: string }, name: { type: string }, companyCode: { type: string } }
    User:
      type: object
      required: [id, name, email, role]
      properties: { id: { type: string }, name: { type: string }, email: { type: string, format: email }, role: { type: string, enum: [SUPER_ADMIN, COMPANY_ADMIN, HR, FINANCE, MANAGER, SUPERVISOR, EMPLOYEE] }, company: { oneOf: [{ $ref: '#/components/schemas/CompanySummary' }, { type: 'null' }] } }
...
    Employee:
      type: object
      required: [id, employeeId, status, user]
      properties: { id: { type: string }, employeeId: { type: string }, department: { type: [string, 'null'] }, designation: { type: [string, 'null'] }, status: { type: string }, user: { $ref: '#/components/schemas/User' } }
...
    ListMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1 }, total: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    ObjectEnvelope: { type: object, required: [data], properties: { data: {}, meta: { $ref: '#/components/schemas/ListMeta' } } }
...